requirements.txt
python-telegram-bot>=22.7
psycopg2-binary>=2.9.0
